Split up dev and prod server in playground
