+++ playground/package.json
@@ -2,10 +2,11 @@
   "name": "stimulus-parser-playground",
   "type": "module",
   "scripts": {
-    "dev": "node server.mjs",
+    "dev": "cross-env NODE_ENV=development node server.dev.mjs",
     "build": "npm run build:client",
-    "build:client": "vite build --ssrManifest --outDir dist/client",
-    "preview": "cross-env NODE_ENV=production node server.mjs"
+    "build:client": "vite build --outDir dist/client",
+    "preview": "cross-env NODE_ENV=production node server.prod.mjs",
+    "serve": "cross-env NODE_ENV=production node server.prod.mjs"
   },
   "dependencies": {
     "@alenaksu/json-viewer": "^2.0.1",
